+++ packages/terra-clinical-header/tests/nightwatch/HeaderTestRoutes.jsx
@@ -12,6 +12,7 @@ import LeftAndRightContentHeader from './LeftAndRightContentHeader';
 import LongTitleAndContentHeader from './LongTitleAndContentHeader';
 import LeftAndRightContentSubheader from './LeftAndRightContentSubheader';
 import NoTitleRightContentHeader from './NoTitleRightContentHeader';
+import LargeLeftAndRightContentsWithLargeTitleHeader from './LargeLeftAndRightContentsWithLargeTitleHeader';
 
 const routes = (
   <div>
@@ -25,6 +26,7 @@ const routes = (
     <Route path="/tests/header-tests/long-title-content" component={LongTitleAndContentHeader} />
     <Route path="/tests/header-tests/left-and-right-content-subheader" component={LeftAndRightContentSubheader} />
     <Route path="/tests/header-tests/no-title-right-content" component={NoTitleRightContentHeader} />
+    <Route path="/tests/header-tests/large-left-and-right-contents-large-title" component={LargeLeftAndRightContentsWithLargeTitleHeader} />
   </div>
 );