+++ packages/terra-clinical-header/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update flex-end styling 
 
 1.5.0 - (September 26, 2017)
 -----------------