+++ packages/terra-clinical-header/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+1.6.0 - (November 15, 2017)
+-----------------
 ### Changed
 * Update flex-end styling